Refactor
pkg/adaptersupport/uber -> pkg/adapterManager
