+++ pkg/adapterManager/BUILD
