+++ pkg/adapterManager/env.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package uber
+package adapterManager
 
 import (
 	"istio.io/mixer/pkg/adapter"
